Abstract all API interaciton to DiningAPI class
